Cleanup the various video buffer
Rename the copy_buf => scan_buf as that is what it really is.
Also rename the copy_ptr => read_ptr, and move all the pointers into a
new struct.

Add a new state machine to the vb2_buffer to make sure we don't
start videocapture in the middle of the ODD field.

Add new defines for the length og a data chunk (0x400 = 1024),
and the lengt of the video data chunk (1020).